// apps/humayro/lib/api/model/cartItemDTO.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Humayro API
 * Humayro e-commerce API
 * OpenAPI spec version: 1.0
 */

export interface CartItemDTO {
  id?: number;
  variantId?: number;
  productId?: number;
  productName?: string;
  colorName?: string;
  colorHex?: string;
  sizeValue?: string;
  imageUrl?: string;
  unitPrice?: number;
  quantity?: number;
  stock?: number;
  subtotal?: number;
}